Remove a test-specific code path from `config`
The module level `astropy.config.configuration._override_config_file`
variable allowed changing the location of the configuration file and was
used in some unit tests. Having test-specific code paths is not a good
idea because the tests should check the actual code. In this case it was
also unnecessary because it is possible to change the configuration file
location using only public API.