Rewrite `EarthLocation` site registry fetching
Astropy maintains a registry of observatory coordinates for the
`EarthLocation.of_site()` method, but in a separate repository. The
`astropy` library defaults to reading a local cached version of the data
and accesses the server only if the cache is empty, so when the data on
the server gets updated `astropy` keeps using the old data. The
`refresh_cache` parameter of `EarthLocation.of_site()` (and
`EarthLocation.get_site_names()`) methods was meant to allow users to
update the cache, but so far it has not worked at all. To fix the bug
the relevant code has been simplified. The separate private
`get_downloaded_sites()` function no longer exists and the undocumented
and untested ability to fetch data from a custom URL has been removed.